@@ -18,27 +18,27 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  cloud_firestore: ^0.16.0
-  connectivity: ^2.0.2
-  cupertino_icons: ^1.0.2
-  file_picker: ^2.1.5+1
-  firebase_auth: ^0.20.0+1
-  firebase_core: ^0.7.0
-  firebase_storage: ^7.0.0
-  flutter_signin_button: ^1.1.0
-  flutter_widget_from_html_core: ^0.5.1+4
-  fluttertoast: ^7.1.6
-  get_it: ^5.0.6
-  getwidget: ^1.2.4
+  cloud_firestore: ^2.0.0
+  connectivity: ^3.0.3
+  cupertino_icons: ^1.0.3
+  file_picker: ^3.0.1
+  firebase_auth: ^1.1.3
+  firebase_core: ^1.1.0
+  firebase_storage: ^8.0.5
+  flutter_signin_button: ^2.0.0
+  fluttertoast: ^8.0.6
+  get: ^4.1.4
+  get_it: ^7.1.3
+  getwidget: ^2.0.2
   google_sign_in: ^4.5.9
-  provider: ^4.3.3
-  shared_preferences: ^0.5.12+4
-  uuid: ^2.2.2
+  provider: ^5.0.0
+  shared_preferences: ^2.0.5
+  uuid: ^3.0.4
 
 dev_dependencies:
   flutter_test:
