@@ -1,37 +1,40 @@
-import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
+import 'package:wasfat_web/firebase/category_service.dart';
 import 'package:wasfat_web/models/food_category.dart';
 
-class CategoryListProvider extends ChangeNotifier {
-  final FirebaseFirestore _firestore;
+class CategoriesProvider extends ChangeNotifier {
+  final CategoryService _categoryService;
 
-  CategoryListProvider(this._firestore);
+  CategoriesProvider(this._categoryService);
 
   List<String> _tempCategoryIdList = <String>[];
-  List<String> get tempCategoryIdsList => _tempCategoryIdList;
+  List<String> get tempCategoryIdsList => this._tempCategoryIdList;
   void addIdToCategoryIdsList(String value) {
-    if (!_tempCategoryIdList.contains(value)) _tempCategoryIdList.add(value);
+    if (!this._tempCategoryIdList.contains(value))
+      this._tempCategoryIdList.add(value);
+    notifyListeners();
+  }
+
+  List<FoodCategory> _categories = [];
+  List<FoodCategory> get categories => this._categories;
+  set categories(List<FoodCategory> value) {
+    this._categories = value;
     notifyListeners();
   }
 
   void removeFromCategoryIdsList(String value) {
-    if (_tempCategoryIdList.contains(value)) _tempCategoryIdList.remove(value);
+    if (this._tempCategoryIdList.contains(value))
+      this._tempCategoryIdList.remove(value);
     notifyListeners();
   }
 
-  void emptyList() {
+  void clearList() {
     _tempCategoryIdList.clear();
     notifyListeners();
   }
 
-  var foodCategories = Map<String, FoodCategory>();
-  Future<void> getFoodCategory() async {
-    final query =
-        await _firestore.collection('food_category').orderBy('priority').get();
-    if (query.docs.isNotEmpty)
-      foodCategories.addEntries(query.docs.map((foodCategory) =>
-          MapEntry<String, FoodCategory>(foodCategory.data()['id'],
-              FoodCategory.fromMap(foodCategory.data()))));
-    notifyListeners();
+  Future<void> getFoodCategories() async {
+    final result = await _categoryService.getCategories();
+    categories = result;
   }
 }
